graceful core restarts
