+++ extensions/intellij/src/main/kotlin/com/github/continuedev/continueintellijextension/continue/CoreMessenger.kt
@@ -180,6 +180,11 @@ class CoreMessenger(private val project: Project, esbuildPath: String, continueC
         Files.setPosixFilePermissions(Paths.get(path), perms)
     }
 
+    private val exitCallbacks: MutableList<() -> Unit> = mutableListOf()
+    fun onDidExit(callback: () -> Unit) {
+        exitCallbacks.add(callback)
+    }
+
     init {
         if (useTcp) {
             try {
@@ -235,6 +240,7 @@ class CoreMessenger(private val project: Project, esbuildPath: String, continueC
             reader = BufferedReader(InputStreamReader(inputStream, StandardCharsets.UTF_8))
 
             process!!.onExit().thenRun {
+                exitCallbacks.forEach { it() }
                 var err = process?.errorStream?.bufferedReader()?.readText()?.trim()
                 if (err != null) {
                     // There are often "⚡️Done in Xms" messages, and we want everything after the last one
@@ -253,6 +259,11 @@ class CoreMessenger(private val project: Project, esbuildPath: String, continueC
                 telemetryService.capture("jetbrains_core_exit", mapOf(
                     "error" to err
                 ))
+
+                // Clean up all resources
+                writer?.close()
+                reader?.close()
+                process?.destroy()
             }
 
             Thread {
